Empty arrays

+++ src/std/mem.cr
@@ -1,8 +1,8 @@
 module mem where
 
-export proc set(src: u8*, val: i32, n: size_t): void
+export proc set(src: u8*, val: u8, n: size_t): void
 {
         for (let i: size_t = 0; i < n; i += 1) {
-                src[i] = (u8)val;
+                src[i] = val;
         }
 }
